fix casting, and dt is in seconds now

+++ GameObject.h
@@ -19,12 +19,12 @@ public:
 	Vector2 velocity;
 	explicit GameObject(float x = 0, float y = 0) { position = Vector2(x, y); isDeleted = false; }
 
-	virtual void Update(DWORD dt, vector<GameObject*>& coObjects, SceneContext* ctx = nullptr) = 0;
+	virtual void Update(float dt, vector<GameObject*>& coObjects, SceneContext* ctx = nullptr) = 0;
 	virtual void Render() = 0;
 
 	virtual bool IsCollidable() { return true; }
 	virtual bool IsBlocking() { return true; }
-	virtual void OnNoCollision(DWORD dt) {} // Call every collision check but returns no collision
+	virtual void OnNoCollision(float dt) {} // Call every collision check but returns no collision
 	virtual void OnCollisionWith(CollisionEvent* event) {} // Call every collision check with collision data
 	virtual Rect GetBoundingBox() = 0;
